+++ bootique-shiro-web-jwt/src/test/java/io/bootique/shiro/web/jwt/AudienceValidationIT.java
@@ -4,49 +4,60 @@ import io.bootique.BQRuntime;
 import io.bootique.jetty.junit5.JettyTester;
 import io.bootique.junit5.BQApp;
 import io.bootique.junit5.BQTest;
+import jakarta.ws.rs.core.HttpHeaders;
+import jakarta.ws.rs.core.Response;
 import org.junit.jupiter.api.Test;
 
-import java.util.*;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 
 @BQTest
 public class AudienceValidationIT extends ShiroWebJwtModuleIT {
 
     private static final JettyTester jetty = JettyTester.create();
+
     @BQApp
     static final BQRuntime app = runtime(jetty, "classpath:io/bootique/shiro/web/jwt/jwt-audience.yml");
 
-    @Override
-    protected JettyTester jetty() {
-        return jetty;
-    }
-
     @Test
-    public void testWithSingleAudienceInToken() {
-        Map<String, ?> map = rolesMap(List.of("role1", "role2", "role3"));
+    public void singleAudience() {
+        Map<String, ?> map = Map.of("roles", List.of("role1", "role2", "role3"));
         List<String> audience = List.of("aud-1");
-        JettyTester.assertOk(getResponse("private-one", map, audience));
+        JettyTester.assertOk(requestWithToken("/private-one", map, audience));
         audience = List.of("aud-2");
-        JettyTester.assertUnauthorized(getResponse("private-two", map, audience));
+        JettyTester.assertUnauthorized(requestWithToken("/private-two", map, audience));
         audience = List.of("aud-3");
-        JettyTester.assertUnauthorized(getResponse("private-three", map, audience));
+        JettyTester.assertUnauthorized(requestWithToken("/private-three", map, audience));
     }
 
     @Test
-    public void testWithCollectionOfAudienceInToken() {
-        Map<String, ?> map = rolesMap(List.of("role1", "role2", "role3"));
+    public void collectionOfAudiences() {
+        Map<String, ?> map = Map.of("roles", List.of("role1", "role2", "role3"));
         List<String> audience = List.of("aud-1", "aud-2");
-        JettyTester.assertOk(getResponse("private-one", map, audience));
+        JettyTester.assertOk(requestWithToken("/private-one", map, audience));
         audience = List.of("aud-2", "aud-3");
-        JettyTester.assertUnauthorized(getResponse("private-two", map, audience));
+        JettyTester.assertUnauthorized(requestWithToken("/private-two", map, audience));
         audience = List.of("aud-1", "aud-3");
-        JettyTester.assertOk(getResponse("private-three", map, audience));
+        JettyTester.assertOk(requestWithToken("/private-three", map, audience));
     }
 
     @Test
-    public void testWithoutAudienceInToken() {
-        Map<String, ?> map = rolesMap(List.of("role1", "role2", "role3"));
-        JettyTester.assertUnauthorized(getResponse("private-one", map, Collections.emptyList()));
-        JettyTester.assertUnauthorized(getResponse("private-two", map, null));
+    public void noAudience() {
+        Map<String, ?> map = Map.of("roles", List.of("role1", "role2", "role3"));
+        JettyTester.assertUnauthorized(requestWithToken("/private-one", map, Collections.emptyList()));
+        JettyTester.assertUnauthorized(requestWithToken("/private-two", map, null));
+    }
+
+    private Response requestWithToken(String path, Map<String, ?> rolesClaim, List<String> audience) {
+
+        String token = WebJwtTests.jwt(rolesClaim, audience, null);
+
+        return jetty.getTarget()
+                .path(path)
+                .request()
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
+                .get();
     }
 }