+++ bootique-shiro-web-jwt/src/test/java/io/bootique/shiro/web/jwt/DefaultJwtIT.java
@@ -4,7 +4,11 @@ import io.bootique.BQRuntime;
 import io.bootique.jetty.junit5.JettyTester;
 import io.bootique.junit5.BQApp;
 import io.bootique.junit5.BQTest;
+import jakarta.ws.rs.core.HttpHeaders;
+import jakarta.ws.rs.core.Response;
 import org.junit.jupiter.api.Test;
+
+import java.time.LocalDateTime;
 import java.util.*;
 
 @BQTest
@@ -14,68 +18,78 @@ public class DefaultJwtIT extends ShiroWebJwtModuleIT {
     @BQApp
     static final BQRuntime app = runtime(jetty, "classpath:io/bootique/shiro/web/jwt/jwt-default.yml");
 
-    @Override
-    protected JettyTester jetty() {
-        return jetty;
+    @Test
+    public void publicAccess() {
+        JettyTester.assertOk(requestWithToken("/public", Collections.emptyMap(), null, null)).assertContent("public");
     }
 
     @Test
-    public void testPublicAccess() {
-        JettyTester.assertOk(getResponse("public", Collections.emptyMap())).assertContent("public");
+    public void role1() {
+        Map<String, ?> map = Map.of("roles", List.of("role1"));
+        JettyTester.assertOk(requestWithToken("/private-one", map, null, null)).assertContent("private-one");
+        JettyTester.assertUnauthorized(requestWithToken("/private-two", map, null, null));
+        JettyTester.assertUnauthorized(requestWithToken("/private-three", map, null, null));
     }
 
     @Test
-    public void testRole1() {
-        Map<String, ?> map = rolesMap(List.of("role1"));
-        JettyTester.assertOk(getResponse("private-one", map)).assertContent("private-one");
-        JettyTester.assertUnauthorized(getResponse("private-two", map));
-        JettyTester.assertUnauthorized(getResponse("private-three", map));
+    public void role2() {
+        Map<String, ?> map = Map.of("roles", List.of("role2"));
+        JettyTester.assertUnauthorized(requestWithToken("/private-one", map, null, null));
+        JettyTester.assertOk(requestWithToken("/private-two", map, null, null)).assertContent("private-two");
+        JettyTester.assertUnauthorized(requestWithToken("/private-three", map, null, null));
     }
 
     @Test
-    public void testRole2() {
-        Map<String, ?> map = rolesMap(List.of("role2"));
-        JettyTester.assertUnauthorized(getResponse("private-one", map));
-        JettyTester.assertOk(getResponse("private-two", map)).assertContent("private-two");
-        JettyTester.assertUnauthorized(getResponse("private-three", map));
+    public void role3() {
+        Map<String, ?> map = Map.of("roles", List.of("role3"));
+        JettyTester.assertUnauthorized(requestWithToken("/private-one", map, null, null));
+        JettyTester.assertUnauthorized(requestWithToken("/private-two", map, null, null));
+        JettyTester.assertOk(requestWithToken("/private-three", map, null, null)).assertContent("private-three");
     }
 
     @Test
-    public void testRole3() {
-        Map<String, ?> map = rolesMap(List.of("role3"));
-        JettyTester.assertUnauthorized(getResponse("private-one", map));
-        JettyTester.assertUnauthorized(getResponse("private-two", map));
-        JettyTester.assertOk(getResponse("private-three", map)).assertContent("private-three");
+    public void role1And3() {
+        Map<String, ?> map = Map.of("roles", List.of("role1", "role3"));
+        JettyTester.assertOk(requestWithToken("/private-one", map, null, null)).assertContent("private-one");
+        JettyTester.assertUnauthorized(requestWithToken("/private-two", map, null, null));
+        JettyTester.assertOk(requestWithToken("/private-three", map, null, null)).assertContent("private-three");
     }
 
     @Test
-    public void testRole1And3() {
-        Map<String, ?> map = rolesMap(List.of("role1", "role3"));
-        JettyTester.assertOk(getResponse("private-one", map)).assertContent("private-one");
-        JettyTester.assertUnauthorized(getResponse("private-two", map));
-        JettyTester.assertOk(getResponse("private-three", map)).assertContent("private-three");
+    public void expiration() {
+        Map<String, ?> map = Map.of("roles", List.of("role1"));
+        JettyTester.assertUnauthorized(requestWithToken("/private-one", map, null, LocalDateTime.now().minusSeconds(15)));
     }
 
     @Test
-    public void testExpiration() {
-        Map<String, ?> map = rolesMap(List.of("role1"));
-        JettyTester.assertUnauthorized(getResponse("private-one", map, 15, null));
+    public void noRolesClaim() {
+        Map<String, ?> map = Map.of("no-roles", List.of("role1"));
+        JettyTester.assertUnauthorized(requestWithToken("/private-one", map, null, null));
+        JettyTester.assertUnauthorized(requestWithToken("/private-two", map, null, null));
+        JettyTester.assertUnauthorized(requestWithToken("/private-three", map, null, null));
     }
 
     @Test
-    public void testWithoutRolesClaim() {
-        Map<String, ?> map = rolesMap("no-roles", List.of("role1"));
-        JettyTester.assertUnauthorized(getResponse("private-one", map));
-        JettyTester.assertUnauthorized(getResponse("private-two", map));
-        JettyTester.assertUnauthorized(getResponse("private-three", map));
-    }
-
-    @Test
-    public void testWithoutAudienceValidation() {
-        Map<String, ?> map = rolesMap(List.of("role1"));
+    public void noAudienceValidation() {
+        Map<String, ?> map = Map.of("roles", List.of("role1"));
         List<String> audience = List.of("aud-1", "aud-2");
-        JettyTester.assertOk(getResponse("private-one", map, audience)).assertContent("private-one");
-        JettyTester.assertUnauthorized(getResponse("private-two", map, audience));
-        JettyTester.assertUnauthorized(getResponse("private-three", map, audience));
+        JettyTester.assertOk(requestWithToken("/private-one", map, audience, null)).assertContent("private-one");
+        JettyTester.assertUnauthorized(requestWithToken("/private-two", map, audience, null));
+        JettyTester.assertUnauthorized(requestWithToken("/private-three", map, audience, null));
+    }
+
+    private Response requestWithToken(String path, Map<String, ?> rolesClaim, List<String> audience, LocalDateTime expiresAt) {
+
+        String token = WebJwtTests.jwt(
+                rolesClaim,
+                audience,
+                expiresAt
+        );
+
+        return jetty.getTarget()
+                .path(path)
+                .request()
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
+                .get();
     }
 }