+++ bootique-shiro-web-oidc/src/test/java/io/bootique/shiro/web/oidc/OidcFilterIT.java
@@ -111,7 +111,7 @@ public class OidcFilterIT {
     }
 
     @Test
-    public void authCookie() throws Exception {
+    public void authCookie() {
 
         String authToken = OidTests.jwt(Map.of("roles", List.of("role1")));
         Response r = OidTests.clientNoRedirects()
@@ -164,12 +164,8 @@ public class OidcFilterIT {
             assertEquals("test-client", data.getFirst("client_id"));
             assertEquals("test-password", data.getFirst("client_secret"));
 
-            try {
-                String authToken = OidTests.jwt(Map.of("roles", List.of("role1")));
-                return Response.ok("{\"access_token\":\"" + authToken + "\"}").build();
-            } catch (Exception e) {
-                return Response.serverError().entity("Unable to generate auth token: " + e.getMessage()).build();
-            }
+            String authToken = OidTests.jwt(Map.of("roles", List.of("role1")));
+            return Response.ok("{\"access_token\":\"" + authToken + "\"}").build();
         }
     }
 }